more dtypes removed

+++ package/MDAnalysis/analysis/msd.py
@@ -380,4 +380,4 @@ class EinsteinMSD(AnalysisBase):
         for n in range(self.n_particles):
             self.msds_by_particle[:, n] = tidynamics.msd(
                 positions[:, n, :])
-        self.timeseries = self.msds_by_particle.mean(axis=1, dtype=np.float64)
+        self.timeseries = self.msds_by_particle.mean(axis=1)
